chore: rename PermissionPersister since it does not fit the role of a persister according to PersisterInterface
