+++ src/Core/Framework/App/Lifecycle/AppLifecycle.php
@@ -23,7 +23,6 @@ use Shopware\Core\Framework\App\Event\PostAppDeletedEvent;
 use Shopware\Core\Framework\App\Exception\AppRegistrationException;
 use Shopware\Core\Framework\App\Lifecycle\Parameters\AppInstallParameters;
 use Shopware\Core\Framework\App\Lifecycle\Parameters\AppUpdateParameters;
-use Shopware\Core\Framework\App\Lifecycle\Persister\PermissionPersister;
 use Shopware\Core\Framework\App\Lifecycle\Persister\PersisterInterface;
 use Shopware\Core\Framework\App\Lifecycle\Registration\AppRegistrationService;
 use Shopware\Core\Framework\App\Manifest\Manifest;
@@ -66,7 +65,7 @@ class AppLifecycle extends AbstractAppLifecycle
     public function __construct(
         private readonly iterable $persisters,
         private readonly EntityRepository $appRepository,
-        private readonly PermissionPersister $permissionPersister,
+        private readonly PermissionLifecycleService $permissionLifecycle,
         private readonly EventDispatcherInterface $eventDispatcher,
         private readonly AppRegistrationService $registrationService,
         private readonly AppStateService $appStateService,
@@ -216,7 +215,7 @@ class AppLifecycle extends AbstractAppLifecycle
 
         $this->updateCustomEntities($app, $manifest);
 
-        $this->permissionPersister->updatePrivileges(
+        $this->permissionLifecycle->updatePrivileges(
             $manifest->getPermissions(),
             $id,
             $manifest->validatesPermissions() === false && $parameters->acceptPermissions,
@@ -308,10 +307,10 @@ class AppLifecycle extends AbstractAppLifecycle
                     'id' => $app->getIntegrationId(),
                     'deletedAt' => new \DateTimeImmutable(),
                 ]], $context);
-                $this->permissionPersister->softDeleteRole($app->getAclRoleId());
+                $this->permissionLifecycle->softDeleteRole($app->getAclRoleId());
             } else {
                 $this->integrationRepository->delete([['id' => $app->getIntegrationId()]], $context);
-                $this->permissionPersister->removeRole($app->getAclRoleId());
+                $this->permissionLifecycle->removeRole($app->getAclRoleId());
             }
 
             $this->deleteAclRole($app->getName(), $context);
