+++ src/Core/Framework/DependencyInjection/app.xml
@@ -65,7 +65,7 @@
             <tag name="shopware.app_manifest.validator"/>
         </service>
 
-        <service id="Shopware\Core\Framework\App\Lifecycle\Persister\PermissionPersister">
+        <service id="Shopware\Core\Framework\App\Lifecycle\PermissionLifecycleService">
             <argument type="service" id="Doctrine\DBAL\Connection"/>
             <argument type="service" id="Shopware\Core\Framework\App\Privileges\Privileges"/>
         </service>
@@ -290,7 +290,7 @@
         <service id="Shopware\Core\Framework\App\Lifecycle\AppLifecycle">
             <argument type="tagged_iterator" tag="shopware.app_lifecycle.persister"/>
             <argument type="service" id="app.repository"/>
-            <argument type="service" id="Shopware\Core\Framework\App\Lifecycle\Persister\PermissionPersister"/>
+            <argument type="service" id="Shopware\Core\Framework\App\Lifecycle\PermissionLifecycleService"/>
             <argument type="service" id="event_dispatcher"/>
             <argument type="service" id="Shopware\Core\Framework\App\Lifecycle\Registration\AppRegistrationService"/>
             <argument type="service" id="Shopware\Core\Framework\App\AppStateService"/>
