+++ tests/unit/Core/Framework/App/Lifecycle/PermissionLifecycleServiceTest.php
@@ -1,12 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Unit\Core\Framework\App\Lifecycle\Persister;
+namespace Shopware\Tests\Unit\Core\Framework\App\Lifecycle;
 
 use Doctrine\DBAL\Connection;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Shopware\Core\Framework\App\Lifecycle\Persister\PermissionPersister;
+use Shopware\Core\Framework\App\Lifecycle\PermissionLifecycleService;
 use Shopware\Core\Framework\App\Manifest\Xml\Permission\Permissions;
 use Shopware\Core\Framework\App\Privileges\Privileges;
 use Shopware\Core\Framework\Context;
@@ -15,20 +15,20 @@ use Shopware\Core\Framework\Uuid\Uuid;
 /**
  * @internal
  */
-#[CoversClass(PermissionPersister::class)]
-class PermissionPersisterTest extends TestCase
+#[CoversClass(PermissionLifecycleService::class)]
+class PermissionLifecycleServiceTest extends TestCase
 {
     private Connection&MockObject $connection;
 
     private Privileges&MockObject $permissions;
 
-    private PermissionPersister $persister;
+    private PermissionLifecycleService $service;
 
     protected function setUp(): void
     {
         $this->connection = $this->createMock(Connection::class);
         $this->permissions = $this->createMock(Privileges::class);
-        $this->persister = new PermissionPersister($this->connection, $this->permissions);
+        $this->service = new PermissionLifecycleService($this->connection, $this->permissions);
     }
 
     public function testUpdatePrivilegesAutoAcceptsIfFlagIsSpecified(): void
@@ -42,7 +42,7 @@ class PermissionPersisterTest extends TestCase
             ->method('setPrivileges')
             ->with($appId, ['customer:read', 'customer:update'], $context);
 
-        $this->persister->updatePrivileges($permissions, $appId, true, $context);
+        $this->service->updatePrivileges($permissions, $appId, true, $context);
     }
 
     public function testUpdatePrivilegesDoesNotAutoAcceptIfFlagIsNotSpecified(): void
@@ -56,6 +56,6 @@ class PermissionPersisterTest extends TestCase
             ->method('requestPrivileges')
             ->with($appId, ['customer:read', 'customer:update'], $context);
 
-        $this->persister->updatePrivileges($permissions, $appId, false, Context::createDefaultContext());
+        $this->service->updatePrivileges($permissions, $appId, false, Context::createDefaultContext());
     }
 }
